remove the file of CopulaABC in examples and put the CopulaABC inference method
